webapp/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-storage==1.7.0
